+++ R/bgmfile.r
@@ -13,42 +13,46 @@
 #' @seealso 
 #' See helper functions to convert the bgm tables to `Spatial` objects, \code{\link{boxSpatial}}, 
 #' \code{\link{faceSpatial}}, \code{\link{nodeSpatial}}, \code{\link{boundarySpatial}}, \code{\link{pointSpatial}}
-#' @importFrom dplyr %>% select distinct_ arrange bind_rows bind_cols distinct mutate mutate_ inner_join
-#' @importFrom tibble as_tibble tibble
+#' @importFrom dplyr %>% select arrange bind_rows bind_cols distinct mutate inner_join
 #' @importFrom utils  head type.convert
-#' @importFrom stringr str_split
-#' @importFrom readr read_file
 #' @examples 
 #' library(bgmfiles)
 #' bfile <- sample(bgmfiles(), 1L)
 #' bgm <- bgmfile(bfile)
-#' library(tibble)
-#' bgm
+#' str(bgm)
 bgmfile <- function(x, ...) {
 
   if (nchar(x) < 1) stop("file path is empty string")
   if (!file.exists(x)) mess <- stop(sprintf("no file found '%s'\n", x))
   #system.time(tx <- readLines(x)) 
-  tx <- stringr::str_split(readr::read_file(x), "\n")[[1]]
+  
+  x0 <- readLines(x)
+  if (all(nchar(x0) < 1)) {
+    stop(sprintf("no lines found in file %s", x))
+  }
+  #browser()
+  tx <- unlist(strsplit(x0, "\n")) ##[[1]]
   nch <- length(tx)
   if (nch == 0) {
     stop(sprintf("no lines found in file %s", x))
   }
+  
   ## all indexes
-#  facesInd <- grep("^face", tx)
-  facesInd <- which(stringr::str_detect(tx, "^face"))
- # boxesInd <- grep("^box", tx)
-  boxesInd <- which(stringr::str_detect(tx, "^box"))
-#  bnd_vertInd <- grep("^bnd_vert", tx)
-  bnd_vertInd <- which(stringr::str_detect(tx, "^bnd_vert"))
+  facesInd <- grep("^face", tx)
+#  facesInd <- which(stringr::str_detect(tx, "^face"))
+  boxesInd <- grep("^box", tx)
+ # boxesInd <- which(stringr::str_detect(tx, "^box"))
+  bnd_vertInd <- grep("^bnd_vert", tx)
+#  bnd_vertInd <- which(stringr::str_detect(tx, "^bnd_vert"))
   ## all comments
-#  hashInd <- grep("^#", tx)
- hashInd <- which(stringr::str_detect(tx, "^#")) 
+  hashInd <- grep("^#", tx)
+ #hashInd <- which(stringr::str_detect(tx, "^#")) 
   ## unique starting tokens
   ust <- sort(unique(unlist(lapply(strsplit(tx[-c(facesInd, boxesInd, bnd_vertInd, hashInd)], "\\s+"), "[", 1))))
   ust <- ust[nchar(ust) > 0]
   extra <- lapply(ust, function(x) gsub("\\s+$", "", gsub("^\\s+", "", gsub(x, "", grep(x, tx, value = TRUE)))))
   names(extra) <- ust
+
   ## some (most?) .bgm have PROJ.4 strings without "+" denoting arguments
   extra$projection <- fixproj(extra$projection)  # <- sprintf("+%s", gsub(" ", " +", extra["projection"]))
   ## nface is repeated in Guam_utm1.bgm
@@ -73,14 +77,14 @@ bgmfile <- function(x, ...) {
   ## we only need boxverts for non-face boxes (boundary faces), but use to check data sense
   boxverts <- dplyr::bind_rows(lapply(seq_along(boxes0), function(xa) {aa <- boxes0[[xa]]$verts; .bx0 = rep(xa - 1, nrow(boxes0[[xa]]$verts)); aa$.bx0 <- .bx0; aa}))
   boxes<- dplyr::bind_rows( 
-                  lapply(boxes0, function(a) dplyr::bind_cols(tibble::as_tibble(a[["meta"]]), 
-                                                              tibble::as_tibble(a[c("insideX", "insideY", ".bx0")]))))
+                  lapply(boxes0, function(a) dplyr::bind_cols(.as_tibble(a[["meta"]]), 
+                                                              .as_tibble(a[c("insideX", "insideY", ".bx0")]))))
   ## ibox/iface are the component faces, and ibox the neighbouring box (.bx0 is the box we belong to!)
   facesXboxes <- dplyr::bind_rows(lapply(boxes0, "[[", "faces"), .id = ".bx0") %>% 
-    dplyr::mutate_(.bx0 = quote(as.numeric(.bx0) - 1))
-  
+    dplyr::mutate(.bx0 = as.numeric(.data$.bx0) - 1)
+
   bnd_verts <- do.call(rbind, lapply(strsplit(tx[bnd_vertInd], "\\s+"), function(x) as.numeric(x[-1])))
-  boundaryverts <- tibble::tibble(x = bnd_verts[,1], y = bnd_verts[,2], bndvert = seq(nrow(bnd_verts)))
+  boundaryverts <- .mk_tibble(x = bnd_verts[,1], y = bnd_verts[,2], bndvert = seq(nrow(bnd_verts)))
   
   for (i in seq(ncol(boxes))) {
     if (is.character(boxes[[i]])) {
@@ -99,11 +103,13 @@ bgmfile <- function(x, ...) {
   
   
   ## I think bnd_verts already all included in box_verts
-  vertices <- bind_rows(faceverts[, c("x", "y")], boxverts[, c("x", "y")], boundaryverts[, c("x", "y")]) %>% distinct_() %>% dplyr::arrange_("x", "y") %>% mutate(.vx0 = row_number())
+  vertices <- bind_rows(faceverts[, c("x", "y")], boxverts[, c("x", "y")], boundaryverts[, c("x", "y")]) %>% 
+    distinct() %>% dplyr::arrange(.data$x, .data$y) %>% mutate(.vx0 = row_number())
   
-  facesXverts <- faceverts %>% mutate(.p0 = rep(1:2, length = nrow(faceverts)))  %>% inner_join(vertices, c("x" = "x", "y" = "y")) %>% dplyr::select_(quote(-x), quote(-y))
+  facesXverts <- faceverts %>% mutate(.p0 = rep(1:2, length = nrow(faceverts)))  %>% inner_join(vertices, c("x" = "x", "y" = "y")) %>% 
+    dplyr::select(-.data$x, -.data$y)
   
-  boxesXverts <- boxverts %>% inner_join(vertices, c("x" = "x", "y" = "y")) %>% dplyr::select_(quote(-x), quote(-y))
+  boxesXverts <- boxverts %>% inner_join(vertices, c("x" = "x", "y" = "y")) %>% dplyr::select(-.data$x, -.data$y)
   
   # allverts <- allverts %>% select(x, y)
   list(vertices = vertices, facesXverts = facesXverts, faces = faces, facesXboxes = facesXboxes, boxesXverts = boxesXverts, boxes = boxes, boundaryvertices = boundaryverts, extra = extra)
@@ -115,14 +121,4 @@ bgmfile <- function(x, ...) {
 read_bgm <- function(x, ...) {
   bgmfile(x, ...)
 }
-# #' @importFrom dplyr select_
-# box2pslg <- function(x) {
-#   x <- head(x$verts, -1) %>% dplyr::select_("x", "y") %>% as.matrix
-#   RTriangle::pslg(x, S = segmaker(x))
-# }
-# segmaker <- function(x) {
-#   on.exit(options(op))
-#   op <- options(warn = -1)
-#   matrix(seq(nrow(x)), nrow = nrow(x) + 1, ncol  = 2)[seq(nrow(x)), ]
-# }